refactor: OnClickShoppingCart 로직 분리
