+++ app/src/main/java/woowacourse/shopping/view/cart/OnClickNavigateShoppingCart.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.view.cart
+
+interface OnClickNavigateShoppingCart {
+    fun clickBack()
+    fun clickOrder()
+    fun clickCartItem(productId: Long)
+}
