+++ app/src/main/java/woowacourse/shopping/view/cart/OnClickShoppingCart.kt
@@ -3,15 +3,10 @@ package woowacourse.shopping.view.cart
 import woowacourse.shopping.domain.model.CartItem
 
 interface OnClickShoppingCart {
-    fun clickBack()
-
-    fun clickCartItem(productId: Long)
 
     fun clickRemoveCartItem(cartItem: CartItem)
 
     fun clickCheckBox(cartItem: CartItem)
-
-    fun clickOrder()
 
     fun clickCheckAll()
 }
