+++ app/src/main/res/layout/item_product.xml
@@ -53,6 +53,7 @@
             android:layout_height="8dp" />
 
         <TextView
+            android:id="@+id/product_name"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="6dp"
